refactor: address GitHub adapter review
