+++ src/common/get-github-adapter.ts
@@ -1,13 +1,11 @@
 import process from 'node:process'
 import * as core from '@actions/core'
 import { context } from '@actions/github'
-// biome-ignore lint/correctness/useImportExtensions: this is a workspace package import.
 import type { Repository } from '@release-drafter/core'
 import {
   createGitHubAdapter,
   type GitHubAdapter,
   type GitHubOctokit,
-  // biome-ignore lint/correctness/useImportExtensions: this is a workspace package import.
 } from '@release-drafter/github-adapter'
 
 export const getRepository = (): Repository => ({
@@ -16,7 +14,9 @@ export const getRepository = (): Repository => ({
   serverUrl: process.env.GITHUB_SERVER_URL ?? 'https://github.com',
 })
 
-export const getGitHubAdapter = (octokit?: GitHubOctokit): GitHubAdapter =>
+let defaultAdapter: GitHubAdapter | undefined
+
+const createAdapter = (octokit?: GitHubOctokit): GitHubAdapter =>
   createGitHubAdapter({
     token: process.env.GITHUB_TOKEN ?? (octokit ? 'injected-client' : ''),
     serverUrl: process.env.GITHUB_SERVER_URL,
@@ -40,3 +40,18 @@ export const getGitHubAdapter = (octokit?: GitHubOctokit): GitHubAdapter =>
         }
       : {}),
   })
+
+/**
+ * Returns the adapter shared by one Action process. An injected Octokit client
+ * still gets its own adapter unless it belongs to the shared instance.
+ */
+export const getGitHubAdapter = (octokit?: GitHubOctokit): GitHubAdapter => {
+  if (octokit) {
+    if (defaultAdapter && defaultAdapter.octokit === octokit)
+      return defaultAdapter
+    return createAdapter(octokit)
+  }
+
+  if (!defaultAdapter) defaultAdapter = createAdapter()
+  return defaultAdapter
+}
